Restore some logging
