+++ lib/cli.js
@@ -1,6 +1,6 @@
 #!/usr/bin/env node
 
-const { UserInterface } = require('./cli/base');
+const { StandardUserInterface } = require('./cli/ui/standard');
 
 // const SearchCommand = require('./cli/search');
 const WakeCommand = require('./cli/wake');
@@ -19,9 +19,11 @@ async function exec(ui, CmdConstructor, ...args) {
 }
 
 async function main() {
-    const ui = new UserInterface({
+    const options = {
         // timeout: 10000,
-    });
+    };
+
+    const ui = new StandardUserInterface(options);
 
     // TODO extract options, figure out which command to run, etc.
     await exec(ui, WakeCommand);
