+++ lib/cli/base.js
@@ -1,54 +1,17 @@
-const readline = require('readline');
 const { Detector, Device, Socket } = require('../../');
 
-/**
- * Default user interface type that logs things directly; tests could
- * use the same interface and implement differently, as could a JSON
- * output mode.
- */
-class UserInterface {
+function setupLogging(ui, d) {
+    d.on('device-notified', (device) => {
+        ui.logEvent('WAKEUP sent to device...', device.address);
+    });
 
-    constructor(options) {
-        if (!options) throw new Error('`options` is required');
+    d.on('logging-in', () => {
+        ui.logEvent('Logging in...');
+    });
 
-        this.options = options;
-        this.detectOpts = options;
-    }
-
-    delayMillis(millis) {
-        return new Promise((resolve) => {
-            setTimeout(resolve, millis);
-        });
-    }
-
-    /* eslint-disable no-console */
-
-    logError(e, ...args) {
-        console.error(e, ...args);
-    }
-
-    logResult(result) {
-        if (typeof result === 'string') {
-            console.log(result);
-        } else {
-            console.log(JSON.stringify(result, null, 2));
-        }
-    }
-
-    /* eslint-enable no-console */
-
-    async prompt(prompt) {
-        return new Promise((resolve) => {
-            readline.createInterface({
-                input: process.stdin,
-                output: process.stdout,
-            }).question(prompt, resolve);
-        });
-    }
-
-    exitWith(code) {
-        process.exit(code);
-    }
+    d.on('sent-key', (k) => {
+        ui.logEvent('Sent key', k);
+    });
 }
 
 class CommandOnDevice {
@@ -117,7 +80,7 @@ class CommandOnDevice {
             this._requestCredentials(ui, d);
         });
 
-        // TODO logging?
+        setupLogging(ui, d);
 
         return d;
     }
@@ -277,6 +240,5 @@ class CommandOnDevice {
 }
 
 module.exports = {
-    UserInterface,
     CommandOnDevice,
 };
